+++ src/components/EditionCard/UpBlurb.tsx
@@ -9,7 +9,7 @@ const UpBlurb: React.FC<{ publishers: Agent[] }> = ({ publishers }) => {
     <Flex
       alignItems={{ base: "start", md: "center" }}
       marginTop={{ base: "s", md: "m" }}
-      marginBottom={{ base: "s", md: null }}
+      marginBottom={{ base: "s", lg: "0" }}
     >
       <Icon name="errorOutline" size="small" />
       <Text size="caption" noSpace marginLeft="xxs">