+++ src/lib/api/SearchApi.ts
@@ -6,6 +6,8 @@ import { toLocationQuery } from "~/src/util/apiConversion";
 import { LinkResult } from "~/src/types/LinkQuery";
 import { ApiLanguageResponse } from "~/src/types/LanguagesQuery";
 import { LOGIN_LINK_BASE } from "~/src/constants/links";
+import { NextRouter } from "next/router";
+import { FulfillResult } from "~/src/types/FulfillQuery";
 
 const apiEnv = process.env["APP_ENV"];
 const apiUrl = process.env["API_URL"] || appConfig.api.url[apiEnv];
@@ -122,7 +124,11 @@ export const readFetcher = async (linkId: number) => {
   }
 };
 
-export const fulfillFetcher = async (fulfillUrl, nyplIdentityCookie) => {
+export const fulfillFetcher = async (
+  fulfillUrl: string,
+  nyplIdentityCookie: any,
+  router: NextRouter
+) => {
   const url = new URL(fulfillUrl);
   const res = await fetch(url.toString(), {
     method: "GET",
@@ -131,12 +137,16 @@ export const fulfillFetcher = async (fulfillUrl, nyplIdentityCookie) => {
     },
   });
   if (res.ok) {
-    return res.url;
+    router.push(res.url);
   } else {
     // redirect to the NYPL login page if access token is invalid
     if (res.status === 401) {
-      return LOGIN_LINK_BASE + encodeURIComponent(window.location.href);
+      router.push(LOGIN_LINK_BASE + encodeURIComponent(window.location.href));
     }
-    throw new Error(`Unable to download UP PDF`);
+    if (res.status === 404) {
+      const fulfillResult: FulfillResult = await res.json();
+      return fulfillResult.data;
+    }
   }
+  return undefined;
 };